+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWETA9CMFFVXXANGSY4J6GA1"
+        "01KWEW4Y17WB2GD0YDT4NKPKXX"
     )
     r"""The unique identifier of the data source"""