+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWETA9CJRSD9B2VK1YPEVXDV"
+        "01KWEW4Y16D8ZDJ9WH93JDBWB6"
     )
     r"""The unique identifier of the data source"""