+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWETA9CQ13WJ0ZVA6P10B3BV"
+        "01KWEW4Y1BYPP693VTHA6QK0F7"
     )
     r"""The unique identifier of the data source"""